pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rosinenpicker"]

[project]
name = "rosinenpicker"
dynamic = ["version"]
authors = [
  { name="Johannes Elias", email="joheli@gmx.net" },
]
description = "A package for picking the juciest text morsels out of a pile of documents."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Development Status :: 2 - Pre-Alpha",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Environment :: Console",
]
dependencies = [
  "pandas>=2.2.0",
  "pydantic>=2.6.1",
  "python-docx>=1.1.0",
  "pyyaml>=6.0.1",
  "pdfminer.six>=20231228",
  "sqlalchemy>=2.0.27",
]

[project.urls]
Homepage = "https://github.com/joheli/rosinenpicker"
Issues = "https://github.com/joheli/rosinenpicker/issues"

[tool.hatch.version]
path = "src/rosinenpicker/start.py"

[project.scripts]
rosinenpicker = "rosinenpicker.start:cli"